requirements.txt
fastapi[standard]==0.122.0
uvicorn[standard]==0.38.0
openai==2.8.1
pymongo==4.15.4
langgraph==1.0.4
langchain==1.1.0
langchain-openai==1.1.0
python-dotenv==1.2.1
pycryptodome==3.23.0
requests==2.32.5